Added a working example for how to init
This will break if tried in solo.
It assumes there to be a FlightControl module.
and that there is a fn-ref with arity 2 in pubsub.

The nice thing is that as long controller implements the behaviour of
FlightControl and you pass any fn-ref with arity 2, it will start and
begin its work.